+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@birdwingo/react-native-reanimated-graph": "^1.1.3",
     "@candlefinance/app-icon": "^0.4.5",
+    "@howljs/calendar-kit": "^2.2.1",
     "@notifee/react-native": "^7.8.2",
     "@react-native-async-storage/async-storage": "1.23.1",
     "@react-native-community/datetimepicker": "8.0.1",
